Add files via upload
Python program for portfolio

Main exe isikukoodiKontrollija.exe takes codes from isikukoodid.txt and calculates birthdate and age. 
For some reason function 4 and 7 are not properly working in the main exe, so for now compiled the functions 4 and 7 separatelly.
Function 4 as in zadanie 4 checks codes for validity and puts correct codes in kontrolitudKoodid.txt, that is originally in the folder with the programm. 
Function 7, takes the correct codes and creates an html with a smiley face as per the template isikTemplate.html

Will get back to this project later